@@ -3,16 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@grapecity/wyn-integration": "^7.0.40",
-    "@types/node": "^12.0.0",
-    "@types/react-dom": "^16.9.0",
-    "bootstrap": "^4.5.3",
-    "node-sass": "^6.0.1",
-    "react": "^16.9.0",
-    "react-bootstrap": "^1.0.0-beta.12",
-    "react-dom": "^16.9.0",
-    "react-icons": "^3.11.0",
-    "typescript": "^3.6.2"
+    "@grapecity/wyn-integration": "^8.1.241",
+    "@types/node": "^24.3.1",
+    "@types/react-dom": "^19.1.9",
+    "bootstrap": "^5.3.8",
+    "react": "^19.1.1",
+    "react-bootstrap": "^2.10.10",
+    "react-dom": "^19.1.1",
+    "react-icons": "^5.5.0",
+    "sass": "^1.92.1",
+    "typescript": "^4.9.5"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -36,7 +36,7 @@
     ]
   },
   "devDependencies": {
-    "@types/react": "^16.9.2",
-    "react-scripts": "4.0.3"
+    "@types/react": "^19.1.12",
+    "react-scripts": "5.0.1"
   }
 }
